requirements.txt
Flask==3.0.0
PyPDF2==3.0.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
ollama==0.2.1
tqdm==4.66.1
Werkzeug==3.0.0
supervisor==4.2.5